+++ src/components/searchfield/searchfield.js
@@ -235,6 +235,8 @@ SearchField.prototype = {
       this.id = utils.uniqueId(this.element, COMPONENT_NAME);
     }
 
+    const html = $('html');
+    this.isSafari = html.is('.is-safari');
     this.label = this.element.prev('label, .label');
     this.inlineLabel = this.element.closest('label');
     this.isInlineLabel = this.element.parent().is('.inline');
@@ -901,6 +903,17 @@ SearchField.prototype = {
       list.off(`focus.${this.id}`);
     });
 
+    // Safari Mac document.activeElement cannot be used on buttons because it is not focusable on the Mac OS Safari.
+    if (this.isSafari) {
+      this.xButton.on('mouseover.clearable', () => {
+        this.closeButtonHover = true;
+      });
+    
+      this.xButton.on('mouseleave.clearable', () => {
+        this.closeButtonHover = false;
+      });
+    }
+
     return this;
   },
 
@@ -957,6 +970,11 @@ SearchField.prototype = {
     const wrapperElem = this.wrapper[0];
 
     if (this.settings.tabbable && wrapperElem.contains(active) && $(active).is('button.close')) {
+      return true;
+    }
+
+    // Safari Mac document.activeElement cannot be used on buttons because it is not focusable on the Mac OS Safari.
+    if (this.isSafari && this.closeButtonHover && !$(active).is('button.close')) {
       return true;
     }
 
@@ -1028,7 +1046,7 @@ SearchField.prototype = {
    * @private
    * @returns {void}
    */
-  handleSafeBlur() {
+  handleSafeBlur(e) {
     const self = this;
     function safeBlurHandler() {
       // Do a check for searchfield-specific elements
@@ -2127,6 +2145,13 @@ SearchField.prototype = {
 
     if (this.xButton && this.xButton.length) {
       this.xButton.off(`blur.${this.id}`);
+
+      // Safari Mac document.activeElement cannot be used on buttons because it is not focusable on the Mac OS Safari.
+      if (this.isSafari) {
+        this.xButton.off('mouseover.clearable');
+        this.xButton.off('mouseleave.clearable');
+      }
+
       this.xButton.remove();
     }
 
@@ -2141,8 +2166,8 @@ SearchField.prototype = {
   makeClearable() {
     this.element.clearable({ tabbable: this.settings.tabbable });
     this.wrapper.addClass('has-close-icon-button');
-    this.xButton = this.wrapper.children('.icon.close');
-
+    this.xButton = this.wrapper.children('button.close');
+  
     // Ignoring the close button from tabbing
     if (!this.settings.tabbable) {
       this.xButton[0].setAttribute('tabindex', '-1');